+++ controller/games.js
@@ -3,7 +3,10 @@ const Rooms = require('../models/rooms');
 const Users = require('../models/users');
 const Boards = require('../models/boards');
 const Bugreport = require('../models/bugReports');
-const { gameUserInfo } = require('../lib/games/gameUserInfo')
+const { gameUserInfo } = require('../lib/games/gameUserInfo');
+const { calculatePoint } = require('../lib/games/gamePoint');
+const { winBlackPointShow } = require('../lib/games/winBlackPointShow');
+const { winWhitePointShow } = require('../lib/games/winWhitePointShow');
 
 
 //대기실 => 게임방 입장시 게임방 생성
@@ -44,12 +47,12 @@ const gameCreate = async (req, res) => {
 const gameStart = async (req, res) => {
   try {
     const { gameNum } = req.params;
-    console.log("47,gameStart,req.params:", req.params)
+    console.log("50,gameStart,req.params:", req.params)
     //게임방내 유저 state별 정보
     let gameInfo = await gameUserInfo(gameNum);
     const gameName = await Games.findOne({ gameNum },{ _id:0, gameNum:1, gameName:1 });  
     const findBoardColor = await Rooms.findOne({ roomNum:gameNum }, { _id:0, boardColor:1 }); 
-    console.log("52,gameStart,gameInfo:",gameInfo, "findBoardColor>>", findBoardColor)
+    console.log("55,gameStart,gameInfo:",gameInfo, "findBoardColor>>", findBoardColor);
     gameInfo.push(findBoardColor)
     res.status(200).json({
       gameInfo,
@@ -65,7 +68,6 @@ const gameStart = async (req, res) => {
     });
   }
 };
-
 
 // [버그리폿] 
 const bugReport = async (req, res) => {
@@ -108,7 +110,7 @@ const bugReport = async (req, res) => {
 const gameFinish = async (req, res) => {
   try {
     const { userInfo, gameNum, result } = req.body;
-    console.log("111,gameFinish,req.body:", req.body)
+    console.log("113,gameFinish,req.body:", req.body)
     const id = userInfo.id;
     const point = userInfo.point;
     const state = userInfo.state;
@@ -121,62 +123,37 @@ const gameFinish = async (req, res) => {
         //승Player
         await Users.updateOne({ id:resultId }, { $inc: { 'score.0.win': 1 } });  //승 +1
         await Users.updateOne({ id:resultId }, { $set: { point: point + 200 } });  //포인트 +200
-        await Users.updateOne({ id:resultId }, { $set: { teachingCnt: 0 }});  // 플레이어는 훈수쳇 계산 안됨.
       } else if(resultId !== id) {
         //패Player
-        await Users.updateOne({ id:id }, { $inc: { 'score.1.lose': 1 } });  //패 +1
-        await Users.updateOne({ id:id }, { $set: { point: point - 100 } });  //포인트 -100
-        await Users.updateOne({ id:id }, { $set: { teachingCnt: 0 }}); 
+        await Users.updateOne({ id }, { $inc: { 'score.1.lose': 1 } });  //패 +1
+        await Users.updateOne({ id }, { $set: { point: point - 100 } });  //포인트 -100
       }
     }
     
-    //Observer
-    //훈수채팅 수 
-    const observerTeachingCnt = await Users.findOne({ id }, { _id: 0, teachingCnt: 1 });
-    const thisTeachingCnt = observerTeachingCnt.teachingCnt;  
-    //이긴팀 point
-    const useTeachingPoint = thisTeachingCnt * 10;  //쓴 포인트
-    const chatPoint = useTeachingPoint * 1  //훈수채팅포인트
-    const addPoint = 20  //승리팀 추가 포인트
-    let winGetTeachingPoint = useTeachingPoint + chatPoint + addPoint;  //얻은 포인트
-    // 포인트제한두기- 최대포인트 300
-    if(winGetTeachingPoint >= 300){
-      winGetTeachingPoint = 300
-    }
-    const winTotalPoint = point + winGetTeachingPoint;  //게임후 총 포인트(기존 + 얻은)
-    console.log("147,gameFinish,winTotalPoint:", winTotalPoint)
-    //진팀 point
-    const penalty = 20  //패널티
-    let loseGetTeachingPoint = useTeachingPoint + chatPoint + penalty;
-    // 포인트제한두기- 최소포인트 300
-    if(loseGetTeachingPoint >= 300){
-      loseGetTeachingPoint = 300
-    } 
-    const loseTotalPoint = point - loseGetTeachingPoint;  //게임후 총 포인트(기존 - 얻은)
-    console.log("156,gameFinish,loseTotalPoint:",loseTotalPoint)
-    //whitePlayer 이김
+    //whitePlayer 승
     if (result.state === 'whitePlayer') {
-      //whiteObserver 이김
+      //whiteObserver 승
       if (state === 'whiteObserver') {
-          await Users.updateOne({ id }, { $set: { point: winTotalPoint } });
-      //blackObserver 짐
+        await calculatePoint({ id, isWin }); 
+        
+        //blackObserver 패
       } else if (state === 'blackObserver') {
-          await Users.updateOne({ id }, { $set: { point: loseTotalPoint } });
+        await calculatePoint({ id, isWin }); 
       }
     }
-    
-    //blackPlayer 이김
+    //blackPlayer 승
     if (result.state === 'blackPlayer') {
-      //blackObserver 이김
+      //blackObserver 승
       if (state === 'blackObserver') {
-          await Users.updateOne({ id }, { $set: { point: winTotalPoint } });
-      //whiteObserver 짐
+        await calculatePoint({ id, isWin }); 
+
+        //whiteObserver 패
       } else if (state === 'whiteObserver') {
-          await Users.updateOne({ id }, { $set: { point: loseTotalPoint } });
+        await calculatePoint({ id, isWin }); 
       }
     }
-    const myId = await Users.findOne({id})
-    console.log("179,gameFinish,myId:",myId)
+    const myId = await Users.findOne({ id })
+    console.log("156,gameFinish,myId:",myId)
 
     res.status(200).json({
       ok: true,
@@ -191,149 +168,34 @@ const gameFinish = async (req, res) => {
   }
 };
 
-
 //[결과창]페이지로 들어가자마자
 const gameFinishShow = async (req, res) => {
   try { 
     const { id, gameNum, result } = req.body;
-    console.log('199,결과창show,req.body:', req.body);
+    console.log('175,결과창show,req.body:', req.body);
     
-    const myId = await Users.findOne({id})
-    console.log("202,show,gameFinishShow,myId:",myId)
+    const myId = await Users.findOne({ id })
+    console.log("178,show,gameFinishShow,myId:",myId)
 
     //게임방내 유저 state별 정보
     const gameInfo = await gameUserInfo(gameNum);
     // console.log(",show,게임디비서찾은,gameInfo[0]:",gameInfo[0]); 
 
     const blackP = gameInfo[0].blackTeamPlayer[0]
-    const blackO = gameInfo[0].blackTeamObserver
     const whiteP = gameInfo[0].whiteTeamPlayer[0]
-    const whiteO = gameInfo[0].whiteTeamObserver
 
     //게임승리 player- black
-    let winPlayerArray = [];
-    let losePlayerArray = [];
-    let winObserverArray1 = [];
-    let loseObserverArray1 = [];
     if (result.win === blackP.id) {
-      //블랙플레이어 승 계산
-      const getBPoint = 200  //우승포인트 
-      const existBPoint = blackP.point - getBPoint  //게임전 포인트(게임후 변경된 포인트 가져옴)
-      const totalBPoint = existBPoint + getBPoint  // 게임후 총 포인트
-      const winInfo = {id:blackP.id, usePoint:0, getPoint:getBPoint, existPoint:existBPoint,
-                                          totalPoint:totalBPoint, state:blackP.state };
-      winPlayerArray.push(winInfo);    
-
-      //화이트플레이어 패 계산
-      const getWPoint = 100  //패배 패널티포인트
-      const existWPoint = whiteP.point + getWPoint //게임전 포인트(게임후 변경된 포인트 가져옴)
-      const totalWPoint = existWPoint - getWPoint // 게임후 총 포인트
-      const loseInfo = {id:whiteP.id, usePoint:0, getPoint: - getWPoint, existPoint:existWPoint,
-                                        totalPoint:totalWPoint, state:whiteP.state };
-      losePlayerArray.push(loseInfo);
-      
-      //블랙옵저버 승 계산
-      for(let i=0; i<blackO.length; i++){
-        //이긴팀 포인트 업데이트
-        const usePoint = blackO[i].teachingCnt * 10;  //쓴포인트 
-        const chatPoint = usePoint * 1  //채팅포인트
-        const addPoint = 20; //추가포인트
-        let getPoint = usePoint + chatPoint + addPoint; //얻은포인트
-        if (getPoint >= 300){
-          getPoint = 300
-        }
-
-        const existPoint = blackO[i].point - getPoint //기존포인트
-        const totalPoint = blackO[i].point;  //총포인트
-        const winObserver = { id:blackO[i].id, usePoint:usePoint, getPoint:getPoint, 
-                              existPoint:existPoint, totalPoint:totalPoint, state:blackO[i].state };
-        winObserverArray1.push(winObserver);
-    }
-      
-      //화이트옵저버 패 계산
-      for(let i=0; i<whiteO.length; i++){
-        //진팀 포인트 업데이트
-        const usePoint = whiteO[i].teachingCnt * 10;  //쓴포인트
-        const chatPoint = usePoint * 1  //채팅포인트
-        const penalty = 20; //진팀 패널티
-        let getPoint = usePoint + chatPoint + penalty; //얻은포인트
-        if (getPoint >= 300){
-          getPoint = 300
-        }
-
-        const existPoint = whiteO[i].point + getPoint //기존포인트
-        const totalPoint = whiteO[i].point;  //총포인트
-        const loseObserver = { id:whiteO[i].id, usePoint:usePoint, getPoint: - getPoint, 
-                                existPoint:existPoint, totalPoint:totalPoint, state:whiteO[i].state };
-        loseObserverArray1.push(loseObserver);
-      } 
+      await winBlackPointShow({ gameNum });
     } 
-
+    
     //게임승리 player- white 
-    let winObserverArray2 = [];
-    let loseObserverArray2 = [];
     if (result.win === whiteP.id) {
-      //화이트플레이어 승 계산
-      const getWPoint = 200  //우승포인트 
-      const existWPoint = whiteP.point - getWPoint //게임전 포인트(게임후 변경된 포인트 가져옴)
-      const totalWPoint = existWPoint + getWPoint  // 게임후 총 포인트
-      const winInfo = {id:whiteP.id, usePoint:0, getPoint:getWPoint, existPoint:existWPoint,
-                                          totalPoint:totalWPoint, state:whiteP.state };
-      winPlayerArray.push(winInfo);    
-
-      //블랙플레이어 패 계산
-      const getBPoint = 100  //패배 패널티포인트
-      const existBPoint = blackP.point + getBPoint //게임전 포인트
-      const totalBPoint = existBPoint - getBPoint // 게임후 총 포인트
-      const loseInfo = {id:blackP.id, usePoint:0, getPoint: - getBPoint, existPoint:existBPoint,
-                                        totalPoint:totalBPoint, state:blackP.state };
-      losePlayerArray.push(loseInfo);
-      
-      //화이트옵저버 승 계산
-      for(let i=0; i<whiteO.length; i++){
-        //이긴팀 포인트 업데이트
-        const usePoint = whiteO[i].teachingCnt  * 10;  //쓴포인트
-        const chatPoint = usePoint * 1  //채팅포인트
-        const addPoint = 20; //추가포인트
-        let getPoint = usePoint + chatPoint + addPoint; //얻은포인트
-          if (getPoint >= 300){
-            getPoint = 300
-          }
-
-        const existPoint = whiteO[i].point - getPoint //기존포인트
-        const totalPoint = whiteO[i].point;  //총포인트
-        const winObserver = { id:whiteO[i].id, usePoint:usePoint, getPoint:getPoint,
-                                existPoint:existPoint, totalPoint:totalPoint, state:whiteO[i].state };
-        winObserverArray2.push(winObserver);
-      }  
-
-      //블랙옵저버 패 계산
-      for(let i=0; i<blackO.length; i++){
-        //진팀 포인트 업데이트
-        const usePoint = blackO[i].teachingCnt * 10;  //쓴포인트
-        const chatPoint = usePoint * 1  //채팅포인트
-        const penalty = 20; //진팀 패널티
-        let getPoint = usePoint + chatPoint + penalty;  //얻은포인트
-        if (getPoint >= 300){
-          getPoint = 300
-        }
-
-        const existPoint = blackO[i].point + getPoint //기존포인트
-        const totalPoint = blackO[i].point;  //총포인트
-        const loseObserver = { id:blackO[i].id, usePoint:usePoint, getPoint: - getPoint, 
-                                  existPoint:existPoint, totalPoint:totalPoint, state:blackO[i].state };
-        loseObserverArray2.push(loseObserver);
+      await winWhitePointShow({ gameNum });
       } 
-    }  
-
-    const win = [...winPlayerArray, ...winObserverArray1, ...winObserverArray2];
-    const lose = [...losePlayerArray, ...loseObserverArray1, ...loseObserverArray2];
-    console.log("331,show,win배열 총정보:",win);
-    console.log("332,show,lose배열 총정보:",lose);
 
     //게임방 결과창 나가기 Observer의 teachingCnt, state, connect변경
     const delTeachingCnt = await Users.findOne({ id },{ _id:0, id:1, state:1, teachingCnt:1 });
-    console.log("336,show,티칭카운트:", delTeachingCnt)
     if(delTeachingCnt.state === 'blackObserver' || delTeachingCnt.state === 'whiteObserver')
       await Users.updateOne({ id }, { $set: { teachingCnt: 0, state: 'online', connect: 'endGame' }});
 
@@ -360,7 +222,7 @@ const gameFinishShow = async (req, res) => {
 const gameDelete = async (req, res) => {
   try {
     const { gameNum } = req.params;
-    console.log("363,gameDelete:",gameDelete)
+    console.log(",gameDelete:",gameDelete)
     const existGame = await Games.findOne({ gameNum });
     if (existGame){
       await Rooms.deleteOne({ roomNum: Number(gameNum) });
